requirements.txt
python-telegram-bot==20.3
aiohttp
SQLAlchemy
python-dotenv
